Add files via upload
In this version for every user-initiated interrupt, i.e. a CTRLC, a new timestamped frame will be generated of the detected blue values vs the time of the run. This is useful for the detection of chimera states in the evolution from desync to sync and is crucial for tomography of the complex networked system under study.